+++ src/Janus.Hosting/Authorization/AuthorizationJson.cs
@@ -20,6 +20,7 @@ namespace Janus.Hosting.Authorization;
 [JsonSerializable(typeof(GroupBody))]
 [JsonSerializable(typeof(GroupRemovalBody))]
 [JsonSerializable(typeof(GroupView))]
+[JsonSerializable(typeof(IReadOnlyList<HeldGrantView>))]
 [JsonSerializable(typeof(IReadOnlyList<GroupView>))]
 [JsonSerializable(typeof(MemberBody))]
 [JsonSerializable(typeof(RoleBody))]